+++ src/Contracts/HasBirdMessage.php
@@ -2,10 +2,4 @@
 
 namespace Spits\Bird\Contracts;
 
-use Spits\Bird\Concerns\IsBirdMessage;
-use Spits\Bird\Enums\IdentifierKey;
-use Spits\Bird\Models\Contact;
-
-trait HasBirdMessage
-{
-}
+trait HasBirdMessage {}