+++ src/Messages/WhatsappMessage.php
@@ -3,28 +3,38 @@
 namespace Spits\Bird\Messages;
 
 use Spits\Bird\Enums\ChannelType;
+use Spits\Bird\Support\MessageTemplate;
 
 class WhatsappMessage extends Message
 {
     public ChannelType $viaChannel = ChannelType::WHATSAPP;
 
-    public function __construct(protected string $templateProjectId, protected string $templateVersion, protected string $templateLocale,protected string $receiver, protected array $templateVariables = [])
-    {
-
+    public function __construct(
+        protected string          $receiver,
+        protected ?MessageTemplate $template = null,
+        protected array $body = [],
+    ) {
+        if (is_null($this->template) && empty($this->body)) {
+            throw new \InvalidArgumentException('WhastsappMessage requires either a template or body');
+        }
     }
 
     #[\Override]
     public function toArray(): array
     {
-        $message['receiver']['contacts'] =[
-            ['identifierValue'=>$this->receiver]
-        ] ;
-        $message['template'] = [
-            'projectId' => $this->templateProjectId,
-            'version' => $this->templateVersion,
-            'locale' => $this->templateLocale,
-            'variables' => $this->templateVariables,
+        $message['receiver']['contacts'] = [
+            [
+                'identifierKey' => "phonenumber",
+                'identifierValue' => $this->receiver],
         ];
+
+        if ($this->template) {
+            $message['template'] = $this->template->toArray();
+        }
+
+        if ($this->body) {
+            $message['body'] = $this->body;
+        }
 
         return $message;
     }