+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to `laravel-bird-api` will be documented in this file.
 
+## Unreleased
+
+## [1.1.0] - 2025-08-25
+
+### Added
+- `MessageTemplate` class to help setting up messages that utilise templates created in Bird
+
+### Changed
+- `WhatsappMessage` class now expects either `MessageTemplate` or `body` parameter to be set for sending a message
+
 ## [1.0.0] - 2025-05-19
 
 - Initial release